requirements.txt
ccxt==4.2.25
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
flask==3.0.0
flask-socketio==5.3.6
apscheduler==3.10.4
python-dotenv==1.0.0
python-telegram-bot==20.7
ta==0.11.0
matplotlib==3.8.2
requests==2.31.0
joblib==1.3.2
colorama==0.4.6
tabulate==0.9.0
